Move buildSrc plugin code under rxbinding.project package
