+++ buildSrc/src/main/groovy/com/jakewharton/rxbinding/project/ValidateBindingsTask.groovy
@@ -1,4 +1,4 @@
-package com.jakewharton.rxbindingproject
+package com.jakewharton.rxbinding.project
 
 import com.github.javaparser.JavaParser
 import com.github.javaparser.ast.CompilationUnit
